Move to component directory
