+++ frontend/src/components/Navbar.vue
