+++ frontend/src/components/post/OldestPosts.vue
@@ -15,8 +15,8 @@ import axios from "axios";
 import InfiniteLoading from "vue-infinite-loading";
 
 // components
-import Post from "../layout/Post";
-import Loader from "../Loader";
+import Post from "../post/Post";
+import Loader from "../icons/Loader";
 
 export default {
 	name: "OldestPosts",
