+++ frontend/src/components/post/Post.vue
@@ -16,7 +16,7 @@
 
 <script>
 // components
-import Vote from "../Vote";
+import Vote from "./Vote";
 
 export default {
 	name: "post",
